Add preliminary ARM implementaion
It only supports __gnu_mcount_nc with gcc verion 4.4 or higher on
EABI toolchains for now.

+++ mcount.c
@@ -298,7 +298,7 @@ static int hook_pltgot(void)
 
 		str = elf_strptr(elf, shstr_idx, shdr.sh_name);
 		if (!strcmp(str, ".got.plt")) {
-			unsigned long *got = (void *)shdr.sh_addr;
+			unsigned long *got = (void *)(long)shdr.sh_addr;
 
 			plthook_resolver_addr = got[2];
 			got[2] = (unsigned long)plt_hooker;
